requirements.txt
beautifulsoup4
Pillow==2.5.0
pyyaml
